alert_configurator/requirements.txt
Flask==2.3.2
pip==23.2.1
PyJWT==2.8.0
requests==2.29.0
setuptools==51.0.0
urllib3==1.26.16
pyyaml==6.0.2
